Added ResponseHelper Class, Fixed Code Climate Issues
